added AbstractStudentEntity class for keeping base information of student record added StudentInsertError class for keeping information about invalid records, that could not be saved to database properly, and so they can be repaired via Student entity update minor changes to controller mappings (making them shorter thus cleaner) simplified logic in kotlin ClassService.kt of assigning collective added students to first opened (not full) class added auxiliary map for handling and preventing from saving duplicated students with unique identifier elevated seeding with random grades when devel profile is selected to separate class improved repository method with better query, in result better assignStudentToFirstOpenClass was created, and less code is needed to find first open (not full) class added identifier to student list file added endpoint & logic to correct improper student records database migrations needed for existing postgresql database when application is launched from main application.properties
