+++ src/main/java/com/schoolmodel/model/dto/NewClassDTO.java
@@ -3,10 +3,6 @@ package com.schoolmodel.model.dto;
 public class NewClassDTO {
     private String name;
 
-    public NewClassDTO(String name) {
-        this.name = name;
-    }
-
     public String getName() {
         return name;
     }
