+++ src/main/java/com/schoolmodel/model/entity/Student.java
@@ -1,162 +1,31 @@
 package com.schoolmodel.model.entity;
 
 import jakarta.persistence.*;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import java.time.Period;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Objects;
 
 @Entity
 @Table(name = "student")
 @EntityListeners(AuditingEntityListener.class)
-public class Student {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    private String name;
-    private String surname;
-    @Column(nullable = false, unique = true)
-    private String code;
+public class Student extends AbstractStudentEntity {
     @OneToMany(fetch = FetchType.EAGER)
     @JoinColumn(name = "student_id")
     private List<Grade> studentGrades;
-    private boolean assigned;
-    @Column(nullable = false)
-    private LocalDate birthDate;
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    private LocalDateTime lastUpdatedAt;
-
-    public Student(String name, String surname, String code, LocalDate birthDate, boolean assigned) {
-        this.name = name;
-        this.surname = surname;
-        this.code = code;
-        this.birthDate = birthDate;
-        this.assigned = assigned;
-    }
-
-    public Student() {}
-
-    @Override
-    public String toString() {
-        return "Student{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", surname='" + surname + '\'' +
-                ", code='" + code + '\'' +
-                ", studentGrades=" + studentGrades +
-                ", assigned=" + assigned +
-                '}';
-    }
-
-    public String simpleDisplay() {
-        return "Student{" +
-                "name='" + name + '\'' +
-                ", surname='" + surname + '\'' +
-                ", code='" + code + '\'' +
-                '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Student student = (Student) o;
-        return code.equals(student.code);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name);
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getSurname() {
-        return surname;
-    }
-
-    public void setSurname(String surname) {
-        this.surname = surname;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(String code) {
-        this.code = code;
-    }
 
     public List<Grade> getStudentGrades() {
         return studentGrades;
     }
 
+    public Student(String name, String surname, String identifier, String code, LocalDate birthDate, boolean assigned) {
+        super(name, surname, identifier, code, birthDate, assigned);
+    }
+
+    public Student() {
+    }
+
     public void setStudentGrades(List<Grade> studentGrades) {
         this.studentGrades = studentGrades;
-    }
-
-    public boolean isAssigned() {
-        return assigned;
-    }
-
-    public void setAssigned(boolean assigned) {
-        this.assigned = assigned;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(LocalDateTime createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public LocalDateTime getLastUpdatedAt() {
-        return lastUpdatedAt;
-    }
-
-    public void setLastUpdatedAt(LocalDateTime lastUpdatedAt) {
-        this.lastUpdatedAt = lastUpdatedAt;
-    }
-
-    public LocalDate getBirthDate() {
-        return birthDate;
-    }
-
-    public void setBirthDate(LocalDate birthDate) {
-        this.birthDate = birthDate;
-    }
-
-    public int getAge() {
-        return calculateAge();
-    }
-
-    private int calculateAge() {
-        return Period.between(this.birthDate, LocalDate.now()).getYears();
     }
 }
