+++ src/main/java/com/schoolmodel/model/enums/InsertStatus.java
@@ -1,0 +1,6 @@
+package com.schoolmodel.model.enums;
+
+public enum InsertStatus {
+    SUCCESS,
+    ERROR
+}
